requirements.txt
streamlit
tflite-runtime
opencv-python-headless
Pillow
numpy
